+++ demo/overview.c
@@ -9,6 +9,7 @@ overview(struct nk_context *ctx)
     static int resize = nk_true;
     static int movable = nk_true;
     static int no_scrollbar = nk_false;
+    static int scale_left = nk_false;
     static nk_flags window_flags = 0;
     static int minimizable = nk_true;
 
@@ -23,6 +24,7 @@ overview(struct nk_context *ctx)
     if (resize) window_flags |= NK_WINDOW_SCALABLE;
     if (movable) window_flags |= NK_WINDOW_MOVABLE;
     if (no_scrollbar) window_flags |= NK_WINDOW_NO_SCROLLBAR;
+    if (scale_left) window_flags |= NK_WINDOW_SCALE_LEFT;
     if (minimizable) window_flags |= NK_WINDOW_MINIMIZABLE;
 
     if (nk_begin(ctx, "Overview", nk_rect(10, 10, 400, 600), window_flags))
@@ -84,6 +86,7 @@ overview(struct nk_context *ctx)
             nk_checkbox_label(ctx, "Movable", &movable);
             nk_checkbox_label(ctx, "No Scrollbar", &no_scrollbar);
             nk_checkbox_label(ctx, "Minimizable", &minimizable);
+            nk_checkbox_label(ctx, "Scale Left", &scale_left);
             nk_tree_pop(ctx);
         }